scripts/requirements.txt
# Essential for Kafka Producer/Consumer
kafka-python==2.0.2
Faker==18.10.1

# Data Processing & ML
numpy==1.21.6
pandas==1.3.5
scikit-learn==1.0.2
joblib==1.1.0
scipy==1.7.3

# Spark Dependencies
pyspark==3.2.1
pyarrow==8.0.0

# Cloud Storage (MinIO/S3)
boto3==1.20.50
s3transfer==0.5.0

# General Utilities
python-dotenv==0.21.0
requests==2.26.0
tqdm==4.64.0
